chore: refactor parser
