+++ Makefile
@@ -103,7 +103,7 @@ lint: pre-commit type-check 						## Run all linting
 .PHONY: coverage
 coverage:  											## Run the tests and generate coverage report
 	@echo "=> Running tests with coverage"
-	@$(PDM) run pytest tests --cov appnlib --cov-report html
+	@$(PDM) run pytest tests --cov stac_generator --cov-report html
 
 .PHONY: test
 test:  												## Run the tests
