+++ stac_generator/core/point/schema.py
@@ -1,8 +1,4 @@
-from typing import Any
-
-import pystac
-
-from stac_generator.core.base.schema import HasColumnInfo, ParsedConfig, SourceConfig
+from stac_generator.core.base.schema import HasColumnInfo, SourceConfig
 
 
 class _PointConfig(HasColumnInfo):
@@ -23,28 +19,3 @@ class _PointConfig(HasColumnInfo):
 
 
 class PointConfig(SourceConfig, _PointConfig): ...
-
-
-class ParsedPointConfig(ParsedConfig, _PointConfig):
-    @classmethod
-    def extract_item(cls, item: pystac.Item) -> dict[str, Any]:
-        result = super().extract_item(item)
-        if "X" not in item.properties:
-            raise ValueError(f"Missing X column in properties of item: {item.id}")
-        if "Y" not in item.properties:
-            raise ValueError(f"Missing Y column in properties of item: {item.id}")
-        result.update(
-            {
-                "X": item.properties["X"],
-                "Y": item.properties["Y"],
-                "Z": item.properties.get("Z", None),
-                "T": item.properties.get("T", None),
-                "date_format": item.properties.get("date_format", "ISO8601"),
-                "epsg": item.properties.get("epsg", 4326),
-            }
-        )
-        return result
-
-    @classmethod
-    def from_item(cls, item: pystac.Item) -> "ParsedPointConfig":
-        return cls.model_validate(cls.extract_item(item))
